% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_mhcnuggets_names.R
\name{are_mhcnuggets_names}
\alias{are_mhcnuggets_names}
\title{Determine if an HLA haplotype name follow the name format
that MHCnuggets uses internally}
\usage{
are_mhcnuggets_names(mhcs)
}
\arguments{
\item{mhcs}{the MHC haplotype names}
}
\description{
Determine if an HLA haplotype name follow the name format
that MHCnuggets uses internally
}
\examples{
library(testthat)

expect_true(all(are_mhcnuggets_names(get_mhc_1_haplotypes())))
expect_true(all(are_mhcnuggets_names(get_mhc_2_haplotypes())))
}
