% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ba_models.R
\name{check_ba_models}
\alias{check_ba_models}
\title{Check the \code{ba_models}}
\usage{
check_ba_models(ba_models)
}
\arguments{
\item{ba_models}{Set to TRUE to use a pure BA model}
}
\value{
Nothing.
}
\description{
Check the \code{ba_models}
Will \link{stop} if it is invalid.
}
\examples{
check_ba_models(TRUE)
check_ba_models(FALSE)
}
\author{
Richèl J.C. Bilderbeek
}
