% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_file_exists.R
\name{check_file_exists}
\alias{check_file_exists}
\title{Check if a file exists.}
\usage{
check_file_exists(filename)
}
\arguments{
\item{filename}{name of a file}
}
\value{
TRUE if MHCnuggets is installed
}
\description{
Check if a file exists.
\link{stop} otherwise.
}
\author{
Richèl J.C. Bilderbeek
}
