% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mhc_class.R
\name{check_mhc_class}
\alias{check_mhc_class}
\title{Check the MHC class.}
\usage{
check_mhc_class(mhc_class)
}
\arguments{
\item{mhc_class}{MHC class. Must be \code{I}, \code{II}
or \link{NA}. Use \link{NA} to let the class be deduced automatically}
}
\description{
Check the MHC class. Will \link{stop} if it is invalid.
}
\examples{
library(testthat)

expect_silent(check_mhc_class("I"))
expect_silent(check_mhc_class("II"))
expect_silent(check_mhc_class(NA))
expect_error(
  check_mhc_class("nonsense"),
  "'mhc_class' must be either 'I', 'II' or NA"
)
}
\author{
Richèl J.C. Bilderbeek
}
