% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mhcnuggets_installation.R
\name{check_mhcnuggets_installation}
\alias{check_mhcnuggets_installation}
\title{Check if MHCnuggets is installed.}
\usage{
check_mhcnuggets_installation(
  mhcnuggetsr_folder = get_default_mhcnuggetsr_folder(),
  ormr_folder_name = get_default_orm_folder_name()
)
}
\value{
Nothing.
}
\description{
Check if MHCnuggets is installed. Will \link{stop} if not.
}
\details{
An MHCnuggets installation has two parts:
\enumerate{
\item The installation of the Python package,
for running the code
\item The download of the Python source code,
which allows the use of example files
}
}
\examples{
if (is_mhcnuggets_installed()) {
  check_mhcnuggets_installation()
}
}
\author{
Richèl J.C. Bilderbeek
}
