% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mhcnuggets_installed.R
\name{check_mhcnuggets_installed}
\alias{check_mhcnuggets_installed}
\title{Check if MHCnuggets is installed. Will \link{stop} if not.}
\usage{
check_mhcnuggets_installed()
}
\description{
Check if MHCnuggets is installed. Will \link{stop} if not.
}
\author{
Richèl J.C. Bilderbeek
}
