% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mhcnuggets_options.R
\name{check_mhcnuggets_options}
\alias{check_mhcnuggets_options}
\title{Check the MHCnuggets options. Will \link{stop} if the options are invalid.}
\usage{
check_mhcnuggets_options(mhcnuggets_options)
}
\arguments{
\item{mhcnuggets_options}{options to run MHCnuggets with,
as can be created by \link{create_mhcnuggets_options}.}
}
\description{
Check the MHCnuggets options. Will \link{stop} if the options are invalid.
}
\examples{
library(testthat)

expect_silent(check_mhcnuggets_options(create_test_mhcnuggets_options()))
expect_error(check_mhcnuggets_options("nonsense"))
}
\author{
Richèl J.C. Bilderbeek
}
