% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mhcnuggets_options.R
\name{check_mhcnuggets_options_names}
\alias{check_mhcnuggets_options_names}
\title{Check the names of the elements in an \code{mhcnuggets_options} list.}
\usage{
check_mhcnuggets_options_names(mhcnuggets_options)
}
\arguments{
\item{mhcnuggets_options}{options to run MHCnuggets with,
as can be created by \link{create_mhcnuggets_options}.}
}
\description{
Check the names of the elements in an \code{mhcnuggets_options} list.
Will \link{stop} if an element is missing.
}
\examples{
library(testthat)

expect_silent(
  check_mhcnuggets_options_names(
    create_test_mhcnuggets_options()
  )
)
expect_error(
  check_mhcnuggets_options_names(
    list(something = "nonsense")
  )
)
}
\author{
Richèl J.C. Bilderbeek
}
