% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mhcnuggets_url.R
\name{check_mhcnuggets_url}
\alias{check_mhcnuggets_url}
\title{Check the \code{mhcnuggets_url}}
\usage{
check_mhcnuggets_url(mhcnuggets_url)
}
\arguments{
\item{mhcnuggets_url}{URL to the MHCnuggets GitHub repository}
}
\value{
Nothing.
}
\description{
Check the \code{mhcnuggets_url}.
Will \link{stop} if it is invalid.
}
\examples{
check_mhcnuggets_url("/home/username")
}
\author{
Richèl J.C. Bilderbeek
}
