% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mhcnuggetsr_folder.R
\name{check_mhcnuggetsr_folder}
\alias{check_mhcnuggetsr_folder}
\title{Check the \code{mhcnuggetsr_folder}}
\usage{
check_mhcnuggetsr_folder(mhcnuggetsr_folder)
}
\arguments{
\item{mhcnuggetsr_folder}{folder where \code{mhcnuggetsr} stores its
temporary files. Use \link{get_default_mhcnuggetsr_folder}
for its default value.}
}
\value{
Nothing.
}
\description{
Check the \code{mhcnuggetsr_folder}.
Will \link{stop} if it is invalid.
}
\examples{
check_mhcnuggetsr_folder("/home/username")
}
\author{
Richèl J.C. Bilderbeek
}
