% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_temp_peptides_path.R
\name{create_temp_peptides_path}
\alias{create_temp_peptides_path}
\title{Create a path to a non-existing temporary file}
\usage{
create_temp_peptides_path(fileext = ".fasta")
}
\arguments{
\item{fileext}{file extension}
}
\description{
Create a path to a non-existing temporary file
}
\examples{
create_temp_peptides_path()
}
\author{
Richèl J.C. Bilderbeek
}
