% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_mhcnuggets_options.R
\name{create_test_mhcnuggets_options}
\alias{create_test_mhcnuggets_options}
\title{Create testing options for MHCnuggets}
\usage{
create_test_mhcnuggets_options(
  mhc_class = NA,
  mhc = "HLA-A02:01",
  ba_models = FALSE,
  verbose = FALSE,
  mhcnuggetsr_folder = get_default_mhcnuggets_folder(),
  mhcnuggets_url = get_mhcnuggets_url(),
  folder_name = "deprecated"
)
}
\arguments{
\item{mhc_class}{MHC class. Must be \code{I}, \code{II}
or \link{NA}. Use \link{NA} to let the class be deduced automatically}

\item{mhc}{the MHC haplotype name}

\item{ba_models}{Set to TRUE to use a pure BA model}

\item{verbose}{set to \link{TRUE} for more debug information}

\item{mhcnuggetsr_folder}{folder where \code{mhcnuggetsr} stores its
temporary files. Use \link{get_default_mhcnuggetsr_folder}
for its default value.}

\item{mhcnuggets_url}{URL to the MHCnuggets GitHub repository}

\item{folder_name}{deprecated name for 'mhcnuggetsr_folder'}
}
\description{
Create a set of testing options to run MHCnuggets with.
The most important setting is the use of a specific haplotype.
}
\examples{
if (is_mhcnuggets_installed()) {
  create_test_mhcnuggets_options()
}
}
\author{
Richèl J.C. Bilderbeek
}
