% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downgrade_pip.R
\name{downgrade_pip}
\alias{downgrade_pip}
\title{Downgrade pip.}
\usage{
downgrade_pip(version)
}
\arguments{
\item{version}{pip version}
}
\value{
Nothing
}
\description{
Downgrade pip.
}
\examples{
\dontrun{
  install_pip()
}
}
\author{
Richèl J.C. Bilderbeek
}
