% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitope_predict.R
\name{epitope_predict}
\alias{epitope_predict}
\title{Predict}
\usage{
epitope_predict(class, peptides_path, mhc, ba_models = FALSE)
}
\arguments{
\item{class}{MHC class. Must be \code{I} or \code{II}}

\item{peptides_path}{the path to the peptides}

\item{mhc}{the MHC haplotype}

\item{ba_models}{Set to TRUE to use a pure BA model}
}
\description{
Predict
}
\author{
Richèl J.C. Bilderbeek
}
