% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_mhcnuggets_folder.R
\name{get_default_mhcnuggets_folder}
\alias{get_default_mhcnuggets_folder}
\title{Get the path to the folder where this package installs
\code{MHCnuggets} by default}
\usage{
get_default_mhcnuggets_folder(
  mhcnuggetsr_folder = get_default_mhcnuggetsr_folder()
)
}
\arguments{
\item{mhcnuggetsr_folder}{folder where \code{mhcnuggetsr} stores its
temporary files. Use \link{get_default_mhcnuggetsr_folder}
for its default value.}
}
\value{
the path to the folder where this package installs
`MHCnuggets`` by default
}
\description{
Get the path to the folder where this package installs
\code{MHCnuggets} by default
}
\examples{
get_default_mhcnuggets_folder()
}
\author{
Richèl J.C. Bilderbeek
}
