% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_mhcnuggetsr_folder.R
\name{get_default_mhcnuggetsr_folder}
\alias{get_default_mhcnuggetsr_folder}
\title{Get the path to the folder where this package installs
the required Python packages by default}
\usage{
get_default_mhcnuggetsr_folder()
}
\value{
the path to the folder where this package installs
the required Python packages by default
}
\description{
Get the path to the folder where this package installs
the required Python packages by default
}
\examples{
get_default_mhcnuggetsr_folder()
}
\seealso{
there are two default folders:
\itemize{
\item the default \code{mhcnuggetsr} folder, as can be obtained
with \link{get_default_mhcnuggetsr_folder}.
This is where the Python packages needed are installed,
using the \code{ormr} package: the \code{ormr_folder_name}
equals the \code{mhcnuggetsr_folder}
\item the default \code{MHCnuggets} folder, as can be obtained
with \link{get_default_mhcnuggets_folder}.
This is where the \code{MHCnuggets} source code is cloned to.
}
}
\author{
Richèl J.C. Bilderbeek
}
