% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_ormr_folder_name.R
\name{get_default_ormr_folder_name}
\alias{get_default_ormr_folder_name}
\title{Get the default \code{ormr} folder name}
\usage{
get_default_ormr_folder_name()
}
\value{
the default \code{ormr} folder name
}
\description{
Get the default \code{ormr} folder name
}
\examples{
get_default_ormr_folder_name()
}
\author{
Richèl J.C. Bilderbeek
}
