% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_python_version.R
\name{get_default_python_version}
\alias{get_default_python_version}
\title{Get the default version of Python used}
\usage{
get_default_python_version()
}
\value{
the default version of Python used as a string
}
\description{
Get the default version of Python used
}
\examples{
get_default_python_version()
}
\author{
Richèl J.C. Bilderbeek
}
