% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_filename.R
\name{get_example_filename}
\alias{get_example_filename}
\title{Get the full path to an MHCnuggets example file}
\usage{
get_example_filename(
  filename = "test_peptides.peps",
  mhcnuggetsr_folder = get_default_mhcnuggetsr_folder()
)
}
\arguments{
\item{filename}{name of the example file, without the path}

\item{mhcnuggetsr_folder}{folder where \code{mhcnuggetsr} stores its
temporary files. Use \link{get_default_mhcnuggetsr_folder}
for its default value.}
}
\value{
the full path to an MHCnuggets example file
}
\description{
Get the full path to an MHCnuggets example file
}
\examples{
if (is_mhcnuggets_installed()) {
  get_example_filename("test_peptides.peps")
}
}
\seealso{
use \link{get_example_filenames} to get all MHCnuggets
example filenames
}
\author{
Richèl J.C. Bilderbeek
}
