% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_filenames.R
\name{get_example_filenames}
\alias{get_example_filenames}
\title{Get the full path to all MHCnuggets example files}
\usage{
get_example_filenames(
  mhcnuggetsr_folder = get_default_mhcnuggetsr_folder(),
  ormr_folder_name = get_default_orm_folder_name()
)
}
\arguments{
\item{mhcnuggetsr_folder}{folder where \code{mhcnuggetsr} stores its
temporary files. Use \link{get_default_mhcnuggetsr_folder}
for its default value.}

\item{ormr_folder_name}{folder used by the \code{ormr} package
to install Python packages to.
Use \code{python3} to install Python packages via \code{pip}.
Default value is \code{python3}}
}
\value{
a character vector with all MHCnuggets example files
}
\description{
Get the full path to all MHCnuggets example files
}
\examples{
if (is_mhcnuggets_installed()) {
  get_example_filenames()
}
}
\seealso{
use \link{get_example_filename} to get the full
path to a MHCnuggets example file
}
\author{
Richèl J.C. Bilderbeek
}
