% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_installed_python_packages.R
\name{get_installed_python_packages}
\alias{get_installed_python_packages}
\title{Get the installed  Python packages and their versions}
\usage{
get_installed_python_packages(
  ormr_folder_name = get_default_ormr_folder_name(),
  python_version = get_default_python_version(),
  verbose = FALSE
)
}
\arguments{
\item{ormr_folder_name}{folder used by the \code{ormr} package
to install Python packages to.
Use \code{python3} to install Python packages via \code{pip}.
Default value is \code{python3}}

\item{python_version}{the Python version as a string}

\item{verbose}{set to \link{TRUE} for more debug information}
}
\value{
a \link[tibble]{tibble} with two rows:
\itemize{
\item \code{package}: the package name
\item \code{version}: the package version
\item \code{requirement}: the packages requirements
\item \code{channel}: the package's channel
}
}
\description{
Get the installed  Python packages and their versions
}
\examples{
if (1 == 2) {
  # Local python3
  get_installed_python_packages(ormr_folder_name = "python3")

  # Conda environment
  if (is_conda_installed()) {
    get_installed_python_packages()
  }
}
}
\author{
Richèl J.C. Bilderbeek
}
