% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mhc_2_haplotypes.R
\name{get_mhc_2_haplotypes}
\alias{get_mhc_2_haplotypes}
\title{Get all the MHC-II haplotypes}
\usage{
get_mhc_2_haplotypes(
  mhcnuggetsr_folder = get_default_mhcnuggets_folder(),
  ormr_folder_name = get_default_orm_folder_name()
)
}
\arguments{
\item{mhcnuggetsr_folder}{folder where \code{mhcnuggetsr} stores its
temporary files. Use \link{get_default_mhcnuggetsr_folder}
for its default value.}

\item{ormr_folder_name}{folder used by the \code{ormr} package
to install Python packages to.
Use \code{python3} to install Python packages via \code{pip}.
Default value is \code{python3}}
}
\value{
a character vector with haplotype names in MHCnuggets format
}
\description{
Get all the MHC-II haplotypes that MHCnuggets has been trained upon.
}
\examples{
if (is_mhcnuggets_installed()) {
  get_mhc_2_haplotypes()
}
}
\author{
Richèl J.C. Bilderbeek
}
