% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mhcnuggets_version.R
\name{get_mhcnuggets_version}
\alias{get_mhcnuggets_version}
\title{Get the MHCnuggets version}
\usage{
get_mhcnuggets_version(
  mhcnuggetsr_folder = get_default_mhcnuggetsr_folder(),
  ormr_folder_name = get_default_orm_folder_name()
)
}
\arguments{
\item{mhcnuggetsr_folder}{folder where \code{mhcnuggetsr} stores its
temporary files. Use \link{get_default_mhcnuggetsr_folder}
for its default value.}

\item{ormr_folder_name}{folder used by the \code{ormr} package
to install Python packages to.
Use \code{python3} to install Python packages via \code{pip}.
Default value is \code{python3}}
}
\value{
a string that is a version, for example \code{2.3.2}
}
\description{
Get the MHCnuggets version
}
\examples{
if (is_mhcnuggets_installed()) {
  get_mhcnuggets_version()
}
}
\author{
Richèl J.C. Bilderbeek
}
