% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_python_package_versions.R
\name{get_python_package_versions}
\alias{get_python_package_versions}
\title{Get the version of all Python packages}
\usage{
get_python_package_versions()
}
\description{
Get the version of all Python packages
}
\examples{
library(testthat)

if (rappdirs::app_dir()$os != "win") {
  df <- get_python_package_versions()
  expect_true(tibble::is_tibble(df))
  expect_true("package" \%in\% names(df))
  expect_true("version" \%in\% names(df))
  expect_true(nrow(df) > 0)
  expect_true(!is.factor(df$package))
  expect_true(!is.factor(df$version))
}
}
\author{
Richèl J.C. Bilderbeek
}
