% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trained_mhc_1_haplotypes.R
\name{get_trained_mhc_1_haplotypes}
\alias{get_trained_mhc_1_haplotypes}
\title{Get all the MHC-I haplotypes that have been trained on a model}
\usage{
get_trained_mhc_1_haplotypes(
  folder_name = get_default_mhcnuggets_folder(),
  mhcnuggets_url = get_mhcnuggets_url()
)
}
\arguments{
\item{folder_name}{superfolder of MHCnuggets.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_mhcnuggets_folder}}

\item{mhcnuggets_url}{URL to the MHCnuggets GitHub repository}
}
\description{
Get all the MHC-I haplotypes that have been trained on a model
}
\examples{
library(testthat)

if (is_mhcnuggets_installed()) {
  haplotypes <- get_trained_mhc_1_haplotypes()
  expect_true(length(haplotypes) > 1)
}
}
\author{
Richèl J.C. Bilderbeek
}
