% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trained_mhc_2_haplotypes.R
\name{get_trained_mhc_2_haplotypes}
\alias{get_trained_mhc_2_haplotypes}
\title{Get all the MHC-II haplotypes that have been trained on a model}
\usage{
get_trained_mhc_2_haplotypes(
  mhcnuggetsr_folder = get_default_mhcnuggets_folder(),
  ormr_folder_name = get_default_orm_folder_name()
)
}
\value{
a character vector with haplotype names in MHCnuggets format
}
\description{
Get all the MHC-II haplotypes that have been trained on a model
}
\examples{
if (is_mhcnuggets_installed()) {
  get_trained_mhc_2_haplotypes()
}
}
\author{
Richèl J.C. Bilderbeek
}
