% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_mhcnuggets.R
\name{install_mhcnuggets}
\alias{install_mhcnuggets}
\title{Install the MHCnuggets Python package.}
\usage{
install_mhcnuggets()
}
\description{
Install the MHCnuggets Python package.
}
\author{
Richèl J.C. Bilderbeek
}
