% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pip.R
\name{install_pip}
\alias{install_pip}
\title{Install pip.}
\usage{
install_pip(
  python_script_filename = file.path(tmpdir = rappdirs::user_cache_dir(),
    "temp_install_pip.py")
)
}
\arguments{
\item{python_script_filename}{name of a temporary Python
script file. Will be deleted at the end of the function}
}
\value{
Nothing
}
\description{
Install pip.
}
\examples{
\dontrun{
  install_pip()
}
}
\author{
Richèl J.C. Bilderbeek
}
