% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_mhcnuggets_name.R
\name{is_mhcnuggets_name}
\alias{is_mhcnuggets_name}
\title{Determine if an HLA haplotype name follow the name format
that MHCnuggets uses internally}
\usage{
is_mhcnuggets_name(mhc)
}
\arguments{
\item{mhc}{the MHC haplotype name}
}
\description{
Determine if an HLA haplotype name follow the name format
that MHCnuggets uses internally
}
\examples{
library(testthat)

expect_true(!is_mhcnuggets_name("HLA-A*01:01"))
expect_true(is_mhcnuggets_name("HLA-A01:01"))
}
