% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_mhcnuggets_options.R
\name{is_mhcnuggets_options}
\alias{is_mhcnuggets_options}
\title{Determine if the MHCnuggets options is valid}
\usage{
is_mhcnuggets_options(mhcnuggets_options)
}
\arguments{
\item{mhcnuggets_options}{options to run MHCnuggets with,
as can be created by \link{create_mhcnuggets_options}.}
}
\description{
Determine if the MHCnuggets options is valid
}
\examples{
library(testthat)

expect_true(is_mhcnuggets_options(create_test_mhcnuggets_options()))
expect_false(is_mhcnuggets_options("nonsense"))
}
\author{
Richèl J.C. Bilderbeek
}
