% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_python_package_installed.R
\name{is_python_package_installed}
\alias{is_python_package_installed}
\title{Determine if a Python package is installed}
\usage{
is_python_package_installed(
  package_name,
  ormr_folder_name = mhcnuggetsr::get_default_ormr_folder_name(),
  python_version = mhcnuggetsr::get_default_python_version(),
  verbose = FALSE
)
}
\arguments{
\item{package_name}{the Python package's name}

\item{ormr_folder_name}{folder used by the \code{ormr} package
to install Python packages to.
Use \code{python3} to install Python packages via \code{pip}.
Default value is \code{python3}}

\item{verbose}{set to \link{TRUE} for more debug information}
}
\value{
nothing
}
\description{
Determine if a Python package is installed
}
\examples{
if (beautier::is_on_ci()) {

  is_python_package_installed(
    ormr_folder_name = "python3",
    package_name = "setuptools"
  )
}
}
\author{
Richèl J.C. Bilderbeek
}
