% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhcnuggets_options_to_text.R
\name{mhcnuggets_options_to_text}
\alias{mhcnuggets_options_to_text}
\title{Convert a \code{mhcnuggets_options} to text}
\usage{
mhcnuggets_options_to_text(mhcnuggets_options)
}
\arguments{
\item{mhcnuggets_options}{options to run MHCnuggets with,
as can be created by \link{create_mhcnuggets_options}.}
}
\value{
Nothing.
}
\description{
Convert a \code{mhcnuggets_options} to text
}
\examples{
if (is_mhcnuggets_installed()) {
  mhcnuggets_options_to_text(create_test_mhcnuggets_options())
}
}
\author{
Richèl J.C. Bilderbeek
}
