% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhcnuggetsr_report.R
\name{mhcnuggetsr_report}
\alias{mhcnuggetsr_report}
\title{Create a \link{mhcnuggetsr} report, to be used when reporting bugs}
\usage{
mhcnuggetsr_report(
  mhcnuggetsr_folder = get_default_mhcnuggetsr_folder(),
  ormr_folder_name = get_default_orm_folder_name()
)
}
\arguments{
\item{mhcnuggetsr_folder}{folder where \code{mhcnuggetsr} stores its
temporary files. Use \link{get_default_mhcnuggetsr_folder}
for its default value.}

\item{ormr_folder_name}{folder used by the \code{ormr} package
to install Python packages to.
Use \code{python3} to install Python packages via \code{pip}.
Default value is \code{python3}}
}
\description{
Create a \link{mhcnuggetsr} report, to be used when reporting bugs
}
\examples{
\dontrun{
  mhcnuggetsr_report()
}
}
\author{
Richèl J.C. Bilderbeek
}
