% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhcnuggetsr_self_test.R
\name{mhcnuggetsr_self_test}
\alias{mhcnuggetsr_self_test}
\title{Self-test the package}
\usage{
mhcnuggetsr_self_test(
  folder_name = get_default_mhcnuggets_folder(),
  mhcnuggets_url = get_mhcnuggets_url()
)
}
\arguments{
\item{folder_name}{superfolder of MHCnuggets.
The name of the superfolder is \code{/home/[user_name]/.local/share}
by default, as can be obtained by\link{get_default_mhcnuggets_folder}}

\item{mhcnuggets_url}{URL to the MHCnuggets GitHub repository}
}
\description{
Self-test the package
}
\examples{
library(testthat)

if (is_mhcnuggets_installed()) {

  expect_silent(mhcnuggetsr_self_test())
}
}
\author{
Richèl J.C. Bilderbeek
}
