% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ic50.R
\name{predict_ic50}
\alias{predict_ic50}
\title{Predict}
\usage{
predict_ic50(
  mhc_class,
  peptides,
  mhc,
  ba_models = FALSE,
  folder_name = get_default_mhcnuggets_folder(),
  mhcnuggets_url = get_mhcnuggets_url(),
  peptides_path = tempfile()
)
}
\arguments{
\item{mhc_class}{MHC class. Must be \code{I} or \code{II}}

\item{peptides}{one of more peptide sequences}

\item{mhc}{the MHC haplotype}

\item{ba_models}{Set to TRUE to use a pure BA model}

\item{folder_name}{superfolder of MHCnuggets.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_mhcnuggets_folder}}

\item{mhcnuggets_url}{URL to the MHCnuggets GitHub repository}

\item{peptides_path}{the path to the peptides}
}
\description{
Predict
}
\examples{
library(testthat)

if (is_mhcnuggets_installed()) {

  peptides <- c("AIAACAMLLV", "ALVCYIVMPV", "ALEPRKEIDV")
  mhc_1_haplotype <- "HLA-A02:01"

  df <- predict_ic50(
    mhc_class = "I",
    peptides = peptides,
    mhc = mhc_1_haplotype
  )
  expect_true("peptide" \%in\% names(df))
  expect_true("ic50" \%in\% names(df))
  expect_equal("character", class(df$peptide))
  expect_equal("numeric", class(df$ic50))
}
}
\author{
Richèl J.C. Bilderbeek
}
