% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ic50_from_file.R
\name{predict_ic50_from_file}
\alias{predict_ic50_from_file}
\title{Predict the half maximal inhibitory concentration (aka IC50) (in nM)
for one or more peptides as saved in a file.
Each peptide must be 15 amino acids at
most (use \link{predict_ic50s} to predict the IC50s for longer peptides)}
\usage{
predict_ic50_from_file(mhcnuggets_options, peptides_path)
}
\arguments{
\item{mhcnuggets_options}{options to run MHCnuggets with,
as can be created by \link{create_mhcnuggets_options}.}

\item{peptides_path}{the path to the peptides}
}
\description{
Predict the half maximal inhibitory concentration (aka IC50) (in nM)
for one or more peptides as saved in a file.
Each peptide must be 15 amino acids at
most (use \link{predict_ic50s} to predict the IC50s for longer peptides)
}
\examples{
library(testthat)

if (is_mhcnuggets_installed()) {

  peptides_path <- get_example_filename("test_peptides.peps")
  expect_true(file.exists(peptides_path))

  mhc_1_haplotype <- "HLA-A02:01"
  expect_true(mhc_1_haplotype \%in\% get_trained_mhc_1_haplotypes())

  mhcnuggets_options <- create_mhcnuggets_options(
    mhc_class = "I",
    mhc = mhc_1_haplotype
  )

  df <- predict_ic50_from_file(
    peptides_path = peptides_path,
    mhcnuggets_options = mhcnuggets_options
  )
  expect_true("peptide" \%in\% names(df))
  expect_true("ic50" \%in\% names(df))
  expect_equal("character", class(df$peptide))
  expect_equal("numeric", class(df$ic50))
}
}
\author{
Richèl J.C. Bilderbeek
}
