% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ic50s.R
\name{predict_ic50s}
\alias{predict_ic50s}
\title{Predict IC50s (in nM) of 15-mers of a peptide}
\usage{
predict_ic50s(
  mhc_class,
  peptide,
  mhc,
  ba_models = FALSE,
  folder_name = get_default_mhcnuggets_folder(),
  mhcnuggets_url = get_mhcnuggets_url(),
  peptides_path = tempfile()
)
}
\arguments{
\item{mhc_class}{MHC class. Must be \code{I} or \code{II}}

\item{peptide}{one peptide sequence}

\item{mhc}{the MHC haplotype name}

\item{ba_models}{Set to TRUE to use a pure BA model}

\item{folder_name}{superfolder of MHCnuggets.
The name of the superfolder is \code{/home/[user_name]/.local/share}
by default, as can be obtained by\link{get_default_mhcnuggets_folder}}

\item{mhcnuggets_url}{URL to the MHCnuggets GitHub repository}

\item{peptides_path}{the path to the peptides}
}
\description{
Predict IC50s (in nM) of 15-mers of a peptide
}
