% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_is_mhcnuggets_installed.R
\name{set_is_mhcnuggets_installed}
\alias{set_is_mhcnuggets_installed}
\title{Set the MHCnuggets installation state to the desired one}
\usage{
set_is_mhcnuggets_installed(
  is_installed,
  verbose = "deprecated",
  folder_name = "deprecated",
  mhcnuggets_url = "deprecated"
)
}
\arguments{
\item{is_installed}{the desired installation state. Must
be \link{TRUE} or \link{FALSE}}

\item{verbose}{set to \link{TRUE} for more debug information}

\item{folder_name}{deprecated name for 'mhcnuggetsr_folder'}

\item{mhcnuggets_url}{URL to the MHCnuggets GitHub repository}
}
\value{
Nothing
}
\description{
Set the MHCnuggets installation state to the desired one
}
\author{
Richèl J.C. Bilderbeek
}
