% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_pip_version.R
\name{set_pip_version}
\alias{set_pip_version}
\title{Set the version of pip.}
\usage{
set_pip_version(version)
}
\arguments{
\item{version}{pip version}
}
\value{
Nothing
}
\description{
Set the version of pip to a specific version,
by installing that version.
}
\examples{
\dontrun{
  if (is_pip_installed()) {
    set_pip_version("19.0")
  }
}
}
\seealso{
use \link{upgrade_pip} to set pip to the latest version.
Use \link{downgrade_pip} to set pip to a specific earlier version.
}
\author{
Richèl J.C. Bilderbeek
}
