% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_mhcnuggets_name.R
\name{to_mhcnuggets_name}
\alias{to_mhcnuggets_name}
\title{Convert a standard haplotype name to the MHCnuggets equivalent}
\usage{
to_mhcnuggets_name(mhc)
}
\arguments{
\item{mhc}{the MHC haplotype name}
}
\description{
Convert a standard haplotype name to the MHCnuggets equivalent
}
\examples{
haplotype <- "HLA-A*01:01"
to_mhcnuggets_name(haplotype)
}
\author{
Richèl J.C. Bilderbeek
}
