% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_mhcnuggets_names.R
\name{to_mhcnuggets_names}
\alias{to_mhcnuggets_names}
\title{Convert one or more standard haplotype name to the MHCnuggets names}
\usage{
to_mhcnuggets_names(mhcs)
}
\arguments{
\item{mhcs}{the MHC haplotype names}
}
\value{
the MHCnuggets names for the haplotypes
}
\description{
Convert one or more standard haplotype names to the MHCnuggets names.
Will \link{stop} if this conversion fails.
}
\examples{
to_mhcnuggets_names("HLA-A*01:01")
}
\author{
Richèl J.C. Bilderbeek
}
