% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninstall_pip.R
\name{uninstall_pip}
\alias{uninstall_pip}
\title{Install pip.}
\usage{
uninstall_pip()
}
\value{
Nothing
}
\description{
Install pip.
}
\examples{
\dontrun{
  uninstall_pip()
}
}
\author{
Richèl J.C. Bilderbeek
}
