% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninstalll_mhcnuggets.R
\name{uninstalll_mhcnuggets}
\alias{uninstalll_mhcnuggets}
\title{Uninstall the MHCnuggets Python package.}
\usage{
uninstalll_mhcnuggets(
  folder_name = get_default_mhcnuggets_folder(),
  mhcnuggets_url = get_mhcnuggets_url()
)
}
\arguments{
\item{folder_name}{superfolder of MHCnuggets.
The name of the superfolder is \code{/home/[user_name]/.local/share}
by default, as can be obtained by\link{get_default_mhcnuggets_folder}}

\item{mhcnuggets_url}{URL to the MHCnuggets GitHub repository}
}
\description{
Uninstall the MHCnuggets Python package.
}
\note{
Use triple-l, due to bug upstream
}
\author{
Richèl J.C. Bilderbeek
}
