% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade_pip.R
\name{upgrade_pip}
\alias{upgrade_pip}
\title{Uograde pip.}
\usage{
upgrade_pip()
}
\value{
Nothing
}
\description{
Uograde pip.
}
\examples{
\dontrun{
  install_pip()
}
}
\seealso{
use \link{downgrade_pip} to set pip to an earlier version.
Use \link{set_pip_version} to install a specific version of pip
}
\author{
Richèl J.C. Bilderbeek
}
