% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mhc_class.R
\name{check_mhc_class}
\alias{check_mhc_class}
\title{Check the MHC class.}
\usage{
check_mhc_class(mhc_class)
}
\arguments{
\item{mhc_class}{MHC class. Must be \code{I}, \code{II}
or \link{NA}. Use \link{NA} to let the class be deduced automatically}
}
\value{
Nothing.
}
\description{
Check the MHC class. Will \link{stop} if it is invalid.
}
\examples{
check_mhc_class("I")
check_mhc_class("II")
check_mhc_class(NA)
}
\author{
Richèl J.C. Bilderbeek
}
