% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mhcnuggets_options.R
\name{check_mhcnuggets_options}
\alias{check_mhcnuggets_options}
\title{Check the MHCnuggets options.}
\usage{
check_mhcnuggets_options(mhcnuggets_options)
}
\arguments{
\item{mhcnuggets_options}{options to run MHCnuggets with,
as can be created by \link{create_mhcnuggets_options}.}
}
\description{
Check the MHCnuggets options.
Will \link{stop} if the options are invalid.
}
\note{
an \code{mhcnuggets_options} created
by \link{create_mhcnuggets_options} is always
checked by \link{check_mhcnuggets_options}
}
\examples{
if (is_mhcnuggets_installed()) {
  check_mhcnuggets_options(create_test_mhcnuggets_options())
}
}
\author{
Richèl J.C. Bilderbeek
}
