% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_mhcnuggets_options.R
\name{create_test_mhcnuggets_options}
\alias{create_test_mhcnuggets_options}
\title{Create a set of testing options to run MHCnuggets with}
\usage{
create_test_mhcnuggets_options(
  mhc_class = "I",
  mhc = "HLA-A02:01",
  ba_models = FALSE,
  verbose = FALSE,
  folder_name = get_default_mhcnuggets_folder(),
  mhcnuggets_url = get_mhcnuggets_url()
)
}
\arguments{
\item{mhc_class}{MHC class. Must be \code{I} or \code{II}}

\item{mhc}{the MHC haplotype name}

\item{ba_models}{Set to TRUE to use a pure BA model}

\item{verbose}{set to \link{TRUE} for more debug information}

\item{folder_name}{superfolder of MHCnuggets.
The name of the superfolder is \code{/home/[user_name]/.local/share}
by default, as can be obtained by\link{get_default_mhcnuggets_folder}}

\item{mhcnuggets_url}{URL to the MHCnuggets GitHub repository}
}
\description{
Create a set of testing options to run MHCnuggets with
}
\examples{
library(testthat)

if (is_mhcnuggets_installed()) {
  options <- create_test_mhcnuggets_options()

  expect_true(is_mhcnuggets_options(options))
  expect_silent(check_mhcnuggets_options(options))
}
}
\author{
Richèl J.C. Bilderbeek
}
