% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(
  ba_models,
  folder_name,
  mhc,
  mhc_class,
  mhcnuggets_url,
  peptides,
  peptides_path
)
}
\arguments{
\item{ba_models}{Set to TRUE to use a pure BA model}

\item{folder_name}{superfolder of MHCnuggets.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_mhcnuggets_folder}}

\item{mhc}{the MHC haplotype}

\item{mhc_class}{MHC class. Must be \code{I} or \code{II}}

\item{mhcnuggets_url}{URL to the MHCnuggets GitHub repository}

\item{peptides}{one of more peptide sequences}

\item{peptides_path}{the path to the peptides}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
\code{@noRd}. This is not done, as this will disallow all
functions to find the documentation parameters
}
\author{
Richèl J.C. Bilderbeek
}
