% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit the documentation
of the parameters from.}
\usage{
default_params_doc(
  ba_models,
  mhc,
  mhcs,
  mhc_class,
  mhcnuggets_options,
  mhcnuggets_output_filename,
  mhcnuggets_url,
  mhcnuggetsr_folder,
  n_aas,
  ormr_folder_name,
  peptide,
  peptides,
  peptide_length,
  peptides_path,
  protein_sequence,
  verbose
)
}
\arguments{
\item{ba_models}{Set to TRUE to use a pure BA model}

\item{mhc}{the MHC haplotype name}

\item{mhcs}{the MHC haplotype names}

\item{mhc_class}{MHC class. Must be \code{I}, \code{II}
or \link{NA}. Use \link{NA} to let the class be deduced automatically}

\item{mhcnuggets_options}{options to run MHCnuggets with,
as can be created by \link{create_mhcnuggets_options}.}

\item{mhcnuggets_output_filename}{name of a temporary file where
MHCnuggets writes its output to}

\item{mhcnuggets_url}{URL to the MHCnuggets GitHub repository}

\item{mhcnuggetsr_folder}{folder where \code{mhcnuggetsr} stores its
temporary files. Use \link{get_default_mhcnuggetsr_folder}
for its default value.}

\item{n_aas}{number of amino acids}

\item{ormr_folder_name}{folder used by the \code{ormr} package
to install Python packages to.
Use \code{python3} to install Python packages via \code{pip}.
Default value is \code{python3}}

\item{peptide}{one peptide sequence}

\item{peptides}{one of more peptide sequences}

\item{peptide_length}{length of a peptide, in number of amino acids}

\item{peptides_path}{the path to the peptides}

\item{protein_sequence}{protein sequence, in uppercase,
for example \code{FAMILYVW}}

\item{verbose}{set to \link{TRUE} for more debug information}
}
\description{
This function does nothing. It is intended to inherit the documentation
of the parameters from.
}
\note{
This is an internal function, so it should be marked with
\code{@noRd}. This is not done, as this will disallow all
functions to find the documentation parameters
}
\author{
Richèl J.C. Bilderbeek
}
