% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downgrade_pip.R
\name{downgrade_pip}
\alias{downgrade_pip}
\title{Downgrade pip.}
\usage{
downgrade_pip(version = "9.0.0")
}
\arguments{
\item{version}{pip version}
}
\value{
Nothing
}
\description{
Set the version of pip to a specific earlier version.
}
\examples{
\dontrun{
  if (is_pip_installed()) {
    downgrade_pip()
  }
}
}
\seealso{
use \link{upgrade_pip} to set pip to the latest version.
Use \link{set_pip_version} to install a specific version of pip
}
\author{
Richèl J.C. Bilderbeek
}
