% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitope_predict.R
\name{epitope_predict}
\alias{epitope_predict}
\title{Predict}
\usage{
epitope_predict(
  class,
  peptides_path,
  mhc,
  ba_models = FALSE,
  folder_name = get_default_mhcnuggets_folder(),
  mhcnuggets_url = get_mhcnuggets_url()
)
}
\arguments{
\item{class}{MHC class. Must be \code{I} or \code{II}}

\item{peptides_path}{the path to the peptides}

\item{mhc}{the MHC haplotype}

\item{ba_models}{Set to TRUE to use a pure BA model}

\item{folder_name}{superfolder of MHCnuggets.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_mhcnuggets_folder}}

\item{mhcnuggets_url}{URL to the MHCnuggets GitHub repository}
}
\description{
Predict
}
\author{
Richèl J.C. Bilderbeek
}
