% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_mhcnuggets_folder.R
\name{get_default_mhcnuggets_folder}
\alias{get_default_mhcnuggets_folder}
\title{Get the path to the folder where this package installs
MHCnuggets by default}
\usage{
get_default_mhcnuggets_folder()
}
\value{
the path to the folder where this package installs
MHCnuggets by default
}
\description{
Get the path to the folder where this package installs
MHCnuggets by default
}
\examples{
library(testthat)

if (rappdirs::app_dir()$os == "unix") {
  expect_true(
    grepl(
      "/home/[A-Za-z0-9_]*/.local/share",
      get_default_mhcnuggets_folder()
    )
  )
}
}
\author{
Richèl J.C. Bilderbeek
}
