% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_filenames.R
\name{get_example_filenames}
\alias{get_example_filenames}
\title{Get the full path to all MHCnuggets example files}
\usage{
get_example_filenames(
  folder_name = get_default_mhcnuggets_folder(),
  mhcnuggets_url = get_mhcnuggets_url()
)
}
\arguments{
\item{folder_name}{superfolder of MHCnuggets.
The name of the superfolder is \code{/home/[user_name]/.local/share}
by default, as can be obtained by\link{get_default_mhcnuggets_folder}}

\item{mhcnuggets_url}{URL to the MHCnuggets GitHub repository}
}
\value{
a character vector with all MHCnuggets example files
}
\description{
Get the full path to all MHCnuggets example files
}
\examples{
if (is_mhcnuggets_installed()) {
  get_example_filenames()
}
}
\seealso{
use \link{get_example_filename} to get the full
path to a MHCnuggets example file
}
\author{
Richèl J.C. Bilderbeek
}
