% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mhc_2_haplotypes.R
\name{get_mhc_2_haplotypes}
\alias{get_mhc_2_haplotypes}
\title{Get all the MHC-II haplotypes that have been trained on a model}
\usage{
get_mhc_2_haplotypes(
  folder_name = get_default_mhcnuggets_folder(),
  mhcnuggets_url = get_mhcnuggets_url()
)
}
\arguments{
\item{folder_name}{superfolder of MHCnuggets.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_mhcnuggets_folder}}

\item{mhcnuggets_url}{URL to the MHCnuggets GitHub repository}
}
\description{
Get all the MHC-II haplotypes that have been trained on a model
}
\examples{
library(testthat)

haplotypes <- get_mhc_2_haplotypes()
expect_true(length(haplotypes) > 1)
}
\author{
Richèl J.C. Bilderbeek
}
