% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mhc_2_haplotypes.R
\name{get_mhc_2_haplotypes}
\alias{get_mhc_2_haplotypes}
\title{Get all the MHC-II haplotypes}
\usage{
get_mhc_2_haplotypes(
  folder_name = get_default_mhcnuggets_folder(),
  mhcnuggets_url = get_mhcnuggets_url()
)
}
\arguments{
\item{folder_name}{superfolder of MHCnuggets.
The name of the superfolder is \code{/home/[user_name]/.local/share}
by default, as can be obtained by\link{get_default_mhcnuggets_folder}}

\item{mhcnuggets_url}{URL to the MHCnuggets GitHub repository}
}
\value{
a character vector with haplotype names in MHCnuggets format
}
\description{
Get all the MHC-II haplotypes that MHCnuggets has been trained upon.
}
\examples{
if (is_mhcnuggets_installed()) {
  get_mhc_2_haplotypes()
}
}
\author{
Richèl J.C. Bilderbeek
}
