% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mhcnuggets_url.R
\name{get_mhcnuggets_url}
\alias{get_mhcnuggets_url}
\title{Get the URL of the MHCnuggets source code}
\usage{
get_mhcnuggets_url()
}
\value{
a string that is a URL
}
\description{
Get the URL of the MHCnuggets source code
}
\examples{
get_mhcnuggets_url()
}
\author{
Richèl J.C. Bilderbeek
}
