% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mhcnuggets_version.R
\name{get_mhcnuggets_version}
\alias{get_mhcnuggets_version}
\title{Get the MHCnuggets version}
\usage{
get_mhcnuggets_version()
}
\description{
Get the MHCnuggets version
}
\author{
Richèl J.C. Bilderbeek
}
