% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_python_package_versions.R
\name{get_python_package_versions}
\alias{get_python_package_versions}
\title{Get the version of all Python packages}
\usage{
get_python_package_versions()
}
\description{
Get the version of all Python packages
}
\examples{
if (rappdirs::app_dir()$os != "win" && is_pip_installed()) {
  get_python_package_versions()
}
}
\author{
Richèl J.C. Bilderbeek
}
