% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_mhcnuggets.R
\name{install_mhcnuggets}
\alias{install_mhcnuggets}
\title{Install the MHCnuggets Python package.}
\usage{
install_mhcnuggets(
  folder_name = get_default_mhcnuggets_folder(),
  mhcnuggets_url = get_mhcnuggets_url()
)
}
\arguments{
\item{folder_name}{superfolder of MHCnuggets.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_mhcnuggets_folder}}

\item{mhcnuggets_url}{URL to the MHCnuggets GitHub repository}
}
\description{
Install the MHCnuggets Python package.
}
\examples{
library(testthat)

if (1 == 2 && is_on_ci() && !is_mhcnuggets_installed()) {
  install_mhcnuggets()
  expect_true(is_mhcnuggets_installed())
}
}
\author{
Richèl J.C. Bilderbeek
}
