% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_mhcnuggets.R
\name{install_mhcnuggets}
\alias{install_mhcnuggets}
\title{Install the \code{MHCnuggets} \code{Python} package.}
\usage{
install_mhcnuggets(
  folder_name = get_default_mhcnuggets_folder(),
  mhcnuggets_url = get_mhcnuggets_url(),
  verbose = FALSE
)
}
\arguments{
\item{folder_name}{superfolder of MHCnuggets.
The name of the superfolder is \code{/home/[user_name]/.local/share}
by default, as can be obtained by\link{get_default_mhcnuggets_folder}}

\item{mhcnuggets_url}{URL to the MHCnuggets GitHub repository}

\item{verbose}{set to \link{TRUE} for more debug information}
}
\value{
Nothing
}
\description{
Install the \code{MHCnuggets} \code{Python} package.
}
\note{
Use triple-l, due to bug upstream,
bug report at \url{https://github.com/r-lib/roxygen2/issues/1101}
}
\examples{
library(testthat)

if (is_on_ci() && !is_mhcnuggets_installed()) {
  # Use triple-l, due to bug upstream,
  # bug report at https://github.com/r-lib/roxygen2/issues/1101
  install_mhcnuggets()
  expect_true(is_mhcnuggets_installed())
}
}
\author{
Richèl J.C. Bilderbeek
}
