% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_mhcnuggets.R
\name{install_mhcnuggets}
\alias{install_mhcnuggets}
\title{Install the \code{MHCnuggets} \code{Python} package.}
\usage{
install_mhcnuggets(folder_name = "deprecated", mhcnuggets_url = "deprecated")
}
\arguments{
\item{folder_name}{deprecated name for 'mhcnuggetsr_folder'}

\item{mhcnuggets_url}{URL to the MHCnuggets GitHub repository}
}
\value{
Nothing
}
\description{
Install the \code{MHCnuggets} \code{Python} package.
}
\examples{
\dontrun{
  install_mhcnuggets()
}
}
\author{
Richèl J.C. Bilderbeek
}
