% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_mhcnuggets.R
\name{install_mhcnuggets}
\alias{install_mhcnuggets}
\title{Install the \code{MHCnuggets} \code{Python} package.}
\usage{
install_mhcnuggets(
  folder_name = get_default_mhcnuggets_folder(),
  mhcnuggets_url = get_mhcnuggets_url()
)
}
\arguments{
\item{folder_name}{superfolder of MHCnuggets.
The name of the superfolder is \code{/home/[user_name]/.local/share}
by default, as can be obtained by\link{get_default_mhcnuggets_folder}}

\item{mhcnuggets_url}{URL to the MHCnuggets GitHub repository}
}
\value{
Nothing
}
\description{
Install the \code{MHCnuggets} \code{Python} package.
}
\examples{
\dontrun{
  install_mhcnuggets()
}
}
\author{
Richèl J.C. Bilderbeek
}
