% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_mhcnuggets_installed.R
\name{is_mhcnuggets_installed}
\alias{is_mhcnuggets_installed}
\title{Check if MHCnuggets is installed}
\usage{
is_mhcnuggets_installed(mhcnuggetsr_folder = get_default_mhcnuggetsr_folder())
}
\arguments{
\item{mhcnuggetsr_folder}{folder where \code{mhcnuggetsr} stores its
temporary files. Use \link{get_default_mhcnuggetsr_folder}
for its default value.}
}
\value{
TRUE if MHCnuggets is installed
}
\description{
Check if MHCnuggets is installed
}
\examples{
is_mhcnuggets_installed()
}
\author{
Richèl J.C. Bilderbeek
}
