% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_mhc_nuggets_installed.R
\name{is_mhcnuggets_installed}
\alias{is_mhcnuggets_installed}
\title{Check if MHCnuggets is installed}
\usage{
is_mhcnuggets_installed()
}
\description{
Check if MHCnuggets is installed
}
\author{
Richèl J.C. Bilderbeek
}
