% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_mhcnuggets_name.R
\name{is_mhcnuggets_name}
\alias{is_mhcnuggets_name}
\title{Is this an MHCnuggets name?}
\usage{
is_mhcnuggets_name(mhc)
}
\arguments{
\item{mhc}{the MHC haplotype name}
}
\value{
TRUE if the name follows the MHCnuggets naming convention
}
\description{
Determine if an HLA haplotype name follow the name format
that MHCnuggets uses internally
}
\examples{
# The official name is not the name format used by MHCnuggets
is_mhcnuggets_name("HLA-A*01:01")

# MHCnuggets uses names without the asterisk
is_mhcnuggets_name("HLA-A01:01")
}
