% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_mhcnuggets_options.R
\name{is_mhcnuggets_options}
\alias{is_mhcnuggets_options}
\title{Is this a \code{mhcnuggets_options}?}
\usage{
is_mhcnuggets_options(mhcnuggets_options)
}
\arguments{
\item{mhcnuggets_options}{options to run MHCnuggets with,
as can be created by \link{create_mhcnuggets_options}.}
}
\value{
TRUE if this a valid set of MHCnuggets options
}
\description{
Determine if the MHCnuggets options is valid.
}
\examples{
if (is_mhcnuggets_installed()) {
  is_mhcnuggets_options(create_test_mhcnuggets_options())
}
}
\author{
Richèl J.C. Bilderbeek
}
