% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhcnuggetsr_report.R
\name{mhcnuggetsr_report}
\alias{mhcnuggetsr_report}
\title{Create a \link{mhcnuggetsr} report, to be used when reporting bugs}
\usage{
mhcnuggetsr_report()
}
\description{
Create a \link{mhcnuggetsr} report, to be used when reporting bugs
}
\examples{
library(testthat)

expect_output(mhcnuggetsr_report())
}
\author{
Richèl J.C. Bilderbeek
}
