% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhcnuggetsr_report.R
\name{mhcnuggetsr_report}
\alias{mhcnuggetsr_report}
\title{Create a \link{mhcnuggetsr} report, to be used when reporting bugs}
\usage{
mhcnuggetsr_report(
  folder_name = get_default_mhcnuggets_folder(),
  mhcnuggets_url = get_mhcnuggets_url()
)
}
\arguments{
\item{folder_name}{superfolder of MHCnuggets.
The name of the superfolder is \code{/home/[user_name]/.local/share}
by default, as can be obtained by\link{get_default_mhcnuggets_folder}}

\item{mhcnuggets_url}{URL to the MHCnuggets GitHub repository}
}
\description{
Create a \link{mhcnuggetsr} report, to be used when reporting bugs
}
\examples{
\dontrun{
  mhcnuggetsr_report()
}
}
\author{
Richèl J.C. Bilderbeek
}
