% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ic50_from_file.R
\name{predict_ic50_from_file}
\alias{predict_ic50_from_file}
\title{Predict}
\usage{
predict_ic50_from_file(
  mhc_class,
  peptides_path,
  mhc,
  ba_models = FALSE,
  folder_name = get_default_mhcnuggets_folder(),
  mhcnuggets_url = get_mhcnuggets_url()
)
}
\arguments{
\item{mhc_class}{MHC class. Must be \code{I} or \code{II}}

\item{peptides_path}{the path to the peptides}

\item{mhc}{the MHC haplotype name}

\item{ba_models}{Set to TRUE to use a pure BA model}

\item{folder_name}{superfolder of MHCnuggets.
The name of the superfolder is \code{/home/[user_name]/.local/share}
by default, as can be obtained by\link{get_default_mhcnuggets_folder}}

\item{mhcnuggets_url}{URL to the MHCnuggets GitHub repository}
}
\description{
Predict
}
\examples{
library(testthat)

if (is_mhcnuggets_installed()) {

  peptides_path <- get_example_filename("test_peptides.peps")
  expect_true(file.exists(peptides_path))

  mhc_1_haplotype <- "HLA-A02:01"
  expect_true(mhc_1_haplotype \%in\% get_trained_mhc_1_haplotypes())

  df <- predict_ic50_from_file(
    mhc_class = "I",
    peptides_path = peptides_path,
    mhc = mhc_1_haplotype
  )
  expect_true("peptide" \%in\% names(df))
  expect_true("ic50" \%in\% names(df))
  expect_equal("character", class(df$peptide))
  expect_equal("numeric", class(df$ic50))
}
}
\author{
Richèl J.C. Bilderbeek
}
