% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ic50s.R
\name{predict_ic50s}
\alias{predict_ic50s}
\title{Predict the half maximal inhibitory concentrations (aka IC50s) (in nM)
of all possible n-mers within a peptide}
\usage{
predict_ic50s(
  peptide,
  n_aas,
  mhcnuggets_options,
  peptides_path = create_temp_peptides_path()
)
}
\arguments{
\item{peptide}{one peptide sequence}

\item{n_aas}{number of amino acids}

\item{mhcnuggets_options}{options to run MHCnuggets with,
as can be created by \link{create_mhcnuggets_options}.}

\item{peptides_path}{the path to the peptides}
}
\description{
Predict the half maximal inhibitory concentrations (aka IC50s) (in nM)
of all possible n-mers within a peptide
}
\examples{
library(testthat)

peptide <- "AIAACAMLLVCCCCCC"

mhcnuggets_options <- create_mhcnuggets_options(
  mhc_class = "I",
  mhc = "HLA-A02:01"
)

df <- predict_ic50s(
  peptide = peptide,
  n_aas = 15,
  mhcnuggets_options = mhcnuggets_options
)

expect_true("peptide" \%in\% names(df))
expect_true("ic50" \%in\% names(df))
expect_equal("character", class(df$peptide))
expect_equal("numeric", class(df$ic50))
n_rows_expected <- nchar(peptide) - 15 + 1
expect_equal(n_rows_expected, nrow(df))
}
\author{
Richèl J.C. Bilderbeek
}
