% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ic50s.R
\name{predict_ic50s}
\alias{predict_ic50s}
\title{Predict the half maximal inhibitory concentrations (aka IC50s) (in nM)
of all possible n-mers within a peptide}
\usage{
predict_ic50s(
  peptide,
  n_aas,
  mhcnuggets_options,
  peptides_path = create_temp_peptides_path()
)
}
\arguments{
\item{peptide}{one peptide sequence}

\item{n_aas}{number of amino acids}

\item{mhcnuggets_options}{options to run MHCnuggets with,
as can be created by \link{create_mhcnuggets_options}.}

\item{peptides_path}{the path to the peptides}
}
\description{
Predict the half maximal inhibitory concentrations (aka IC50s) (in nM)
of all possible n-mers within a peptide
}
\examples{
if (is_mhcnuggets_installed()) {
  peptide <- "AIAACAMLLVCCCCCC"

  mhcnuggets_options <- create_mhcnuggets_options(
    mhc = "HLA-A02:01"
  )

  predict_ic50s(
    peptide = peptide,
    n_aas = 15,
    mhcnuggets_options = mhcnuggets_options
  )
}
}
\author{
Richèl J.C. Bilderbeek
}
