% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninstall_mhcnuggets.R
\name{uninstall_mhcnuggets}
\alias{uninstall_mhcnuggets}
\title{Uninstall the MHCnuggets Python package.}
\usage{
uninstall_mhcnuggets(folder_name = "deprecated", mhcnuggets_url = "deprecated")
}
\arguments{
\item{folder_name}{deprecated name for 'mhcnuggetsr_folder'}

\item{mhcnuggets_url}{URL to the MHCnuggets GitHub repository}
}
\description{
Uninstall the MHCnuggets Python package.
}
\author{
Richèl J.C. Bilderbeek
}
